[新增] 生日查询 [新增] 群发功能

+++ esmwiki-common/src/main/java/moe/zr/qqbot/entry/SendMessage.java
@@ -6,6 +6,6 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class SendMessage {
-    Integer groupId = 773891409;
+    Long groupId = 773891409L;
     String message;
 }
